requirements.txt
requests==2.31.0
python-dotenv==1.0.0
adal==1.2.7
tabulate==0.9.0 